chore: move RifRbtcTooltip and DottedUnderlineLabel to avoid cousin imports
